requirements.txt
pycryptodomex=3.17
prettytable=3.7.0